requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
Django==4.1
django-cors-headers==3.13.0
django-import-export==2.8.0
djongo==1.3.6
dnspython==2.2.1
ecdsa==0.18.0
environs==9.5.0
et-xmlfile==1.1.0
filelock==3.8.0
greenlet==1.1.3
idna==3.4
iniconfig==1.1.1
joblib==1.2.0
jwt==1.3.1
MarkupPy==1.14
marshmallow==3.18.0
mongoengine==0.24.2
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.4
platformdirs==2.5.2
playwright==1.26.1
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pyee==8.1.0
PyJWT==2.6.0
pymongo==3.12.1
pyparsing==3.0.9
pytest==7.1.3
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-slugify==6.1.2
python-snappy==0.6.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sqlparse==0.2.4
tablib==3.2.1
text-unidecode==1.3
threadpoolctl==3.1.0
tomli==2.0.1
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.3
websockets==10.1
winkerberos==0.9.0
xlrd==2.0.1
xlwt==1.3.0
